Update mnist data fetch/conversion scripts.
Takes the latest versions from upstream caffe master. I was having
trouble with the previously included versions where they were not
appropriately creating a valid leveldb that could be read by caffe.
Probably the issue was the key "leveldb" instead of "lmdb" but the
slight updates in the upstream caffe versions couldn't be bad to have,
too.